renames parseAvRoute to parseRoute
