+++ ws-nextjs-app/pages/[service]/av-embeds/handleAvRoute.ts
@@ -6,7 +6,7 @@ import certsRequired from '#app/routes/utils/certsRequired';
 import getEnvironment from '#app/routes/utils/getEnvironment';
 import { FetchError } from '#app/models/types/fetch';
 import constructPageFetchUrl from '#app/routes/utils/constructPageFetchUrl';
-import parseAvRoute from '#app/routes/utils/parseAvRoute';
+import parseAvRoute from '#app/routes/utils/parseRoute';
 import filterForBlockType from '#app/lib/utilities/blockHandlers';
 import nodeLogger from '#lib/logger.node';
 import { OK } from '#app/lib/statusCodes.const';
